Refactor braid theme creation now that treat is gone
