--- packages/braid-design-system/lib/themes/BraidTheme.ts
@@ -1,5 +1,0 @@
-import type { BraidTheme as OldBraidTheme } from './makeBraidTheme';
-
-export interface BraidTheme extends OldBraidTheme {
-  vanillaTheme: string;
-}
