+++ packages/braid-design-system/lib/themes/apac/index.ts
@@ -1,5 +1,1 @@
-import vanillaTheme from './apacTheme.css';
-
-export default {
-  vanillaTheme,
-};
+export { default } from './apacTheme.css';
