+++ packages/braid-design-system/lib/themes/occ/index.ts
@@ -1,5 +1,1 @@
-import vanillaTheme from './occTheme.css';
-
-export default {
-  vanillaTheme,
-};
+export { default } from './occTheme.css';
