+++ packages/braid-design-system/lib/themes/seekAnz/index.ts
@@ -1,5 +1,1 @@
-import vanillaTheme from './seekAnzTheme.css';
-
-export default {
-  vanillaTheme,
-};
+export { default } from './seekAnzTheme.css';
