+++ packages/braid-design-system/lib/themes/seekBusiness/index.ts
@@ -1,5 +1,1 @@
-import vanillaTheme from './seekBusinessTheme.css';
-
-export default {
-  vanillaTheme,
-};
+export { default } from './seekBusinessTheme.css';
