+++ packages/braid-design-system/lib/themes/vars.css.ts
@@ -1,6 +1,6 @@
 import { createThemeContract } from '@vanilla-extract/css';
 
-import makeVanillaTheme from './makeVanillaTheme';
+import { makeVanillaTheme } from './makeVanillaTheme';
 import tokens from './docs/tokens';
 
 export const vars = createThemeContract(makeVanillaTheme(tokens));
