+++ packages/braid-design-system/lib/themes/wireframe/index.ts
@@ -1,5 +1,1 @@
-import vanillaTheme from './wireframeTheme.css';
-
-export default {
-  vanillaTheme,
-};
+export { default } from './wireframeTheme.css';
